requirements.txt
django
django-auth-ldap
django-extensions
django-filter
django-resized
pillow
python-dotenv
python-ldap
